generic_food_processing_company_site/requirements.txt
Django==3.2.8
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
requests==2.31.0
psycopg2-binary==2.8.6  # PostgreSQL adapter
gunicorn==20.0.4  # WSGI server for running the app
django-filter==2.4.0
pandas==1.3.3
openpyxl==3.0.9